epsr: explicitely convert from REAL to INT

+++ epsr.f90
@@ -197,7 +197,7 @@ subroutine epsr_basic(r,dvdr,v,vir,na,boxlxyz,njump)
         dz = dz - boxz*dble(nint(onboxz*dz))
         drsq = dx*dx + dy*dy + dz*dz
         sq = sqrt(drsq)
-        bin = sq/(pos(2)-pos(1))
+        bin = int(sq/(pos(2)-pos(1)))
         if (bin .lt. 1000 .and. bin.gt.0) then
            v = v + potOO(bin)
            dfx = frcOO(bin) * dx*dx/drsq
@@ -232,7 +232,7 @@ subroutine epsr_basic(r,dvdr,v,vir,na,boxlxyz,njump)
           dz = dz - boxz*dble(nint(onboxz*dz))
           drsq = dx*dx + dy*dy + dz*dz
           sq = sqrt(drsq)
-          bin = sq/(pos(2)-pos(1))
+          bin = int(sq/(pos(2)-pos(1)))
           if (bin .lt. 1000 .and. bin.gt.0) then
              v = v + potOH(bin)
              dfx = frcOH(bin) * dx*dx/drsq
@@ -265,7 +265,7 @@ subroutine epsr_basic(r,dvdr,v,vir,na,boxlxyz,njump)
           dz = dz - boxz*dble(nint(onboxz*dz))
           drsq = dx*dx + dy*dy + dz*dz
           sq = sqrt(drsq)
-          bin = sq/(pos(2)-pos(1))
+          bin = int(sq/(pos(2)-pos(1)))
           if (bin .lt. 1000 .and. bin.gt.0) then
              v = v + potOH(bin)
              dfx = frcOH(bin) * dx*dx/drsq
@@ -300,7 +300,7 @@ subroutine epsr_basic(r,dvdr,v,vir,na,boxlxyz,njump)
 !!!        dz = dz - boxz*dble(nint(onboxz*dz))
 !!!        drsq = dx*dx + dy*dy + dz*dz
 !!!        sq = sqrt(drsq)
-!!!        bin = sq/(pos(2)-pos(1))
+!!!        bin = int(sq/(pos(2)-pos(1)))
 !!!        if (bin .lt. 1000 .and. bin.gt.0) then
 !!!           v = v + potOO(bin)
 !!!           dfx = frcHH(bin) * dx*dx/drsq
@@ -332,7 +332,7 @@ subroutine epsr_basic(r,dvdr,v,vir,na,boxlxyz,njump)
 !!!        dz = dz - boxz*dble(nint(onboxz*dz))
 !!!        drsq = dx*dx + dy*dy + dz*dz
 !!!        sq = sqrt(drsq)
-!!!        bin = sq/(pos(2)-pos(1))
+!!!        bin = int(sq/(pos(2)-pos(1)))
 !!!        if (bin .lt. 1000 .and. bin.gt.0) then
 !!!           v = v + potOO(bin)
 !!!           dfx = frcHH(bin) * dx*dx/drsq
@@ -367,7 +367,7 @@ subroutine epsr_basic(r,dvdr,v,vir,na,boxlxyz,njump)
           dz = dz - boxz*dble(nint(onboxz*dz))
           drsq = dx*dx + dy*dy + dz*dz
           sq = sqrt(drsq)
-          bin = sq/(pos(2)-pos(1))
+          bin = int(sq/(pos(2)-pos(1)))
           if (bin .lt. 1000 .and. bin.gt.0) then
              v = v + potHH(bin)
              dfx = frcHH(bin) * dx*dx/drsq
@@ -400,7 +400,7 @@ subroutine epsr_basic(r,dvdr,v,vir,na,boxlxyz,njump)
           dz = dz - boxz*dble(nint(onboxz*dz))
           drsq = dx*dx + dy*dy + dz*dz
           sq = sqrt(drsq)
-          bin = sq/(pos(2)-pos(1))
+          bin = int(sq/(pos(2)-pos(1)))
           if (bin .lt. 1000 .and. bin.gt.0) then
              v = v + potHH(bin)
              dfx = frcHH(bin) * dx*dx/drsq
